requirement.txt
flask
bootstrap-flask
flask_wtf
flask_sqlalchemy
openai
alembic